requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cpf-generator==2.1.4
cryptography==41.0.1
deprecation==2.1.0
Django==4.2.2
django-cron==0.6.0
django-crontab==0.7.1
django-extensions==3.2.3
django-ninja==0.22.2
idna==3.4
iniconfig==2.0.0
names==0.3.0
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pydantic==1.10.9
pyOpenSSL==23.2.0
pytest==7.3.2
pytest-django==4.5.2
requests==2.31.0
sqlparse==0.4.4
starkbank==2.20.0
starkbank-ecdsa==2.2.0
starkcore==0.1.1
typing_extensions==4.6.3
urllib3==2.0.3
